Add mocked embedding service for tests
The tests were failing with 503 "Model not loaded" because TestClient
doesn't execute the FastAPI lifespan context by default. Loading the
actual ML model during tests would be slow and resource-intensive.

Changes:
- Created tests/conftest.py with shared fixtures
- Added mock_embedding_service fixture that mocks the model
- Mock returns appropriate number of embeddings based on input
- Moved client and auth_headers fixtures to conftest.py
- Cleaned up test_api.py by removing duplicate fixtures

This makes tests fast and independent of external model downloads.
